About this Repository
This repository consists of the Student Portfolio project that I have completed. In this project, I had made changes to the CSS to personalize the page and make it more "me".  For this course, we had learned how to write html docs and style them utilizing css. We were tasked with not only learning these coding languages but also understanding what it is we were looking at. Through this course, it has culminated in the knowledge base to now be able to build and style simple websites which now provides us with the opportunity to learn new skills to make these sites more dynamic and interactive.